+++ packages/starknet-snap/src/rpcs/display-private-key.ts
