+++ packages/starknet-snap/src/rpcs/index.ts
@@ -1,8 +1,8 @@
-export * from './displayPrivateKey';
-export * from './estimateFee';
-export * from './executeTxn';
-export * from './signMessage';
-export * from './signTransaction';
+export * from './display-private-key';
+export * from './estimate-fee';
+export * from './execute-txn';
+export * from './sign-message';
+export * from './sign-transaction';
 export * from './sign-declare-transaction';
 export * from './verify-signature';
 export * from './switch-network';
