+++ packages/starknet-snap/src/rpcs/sign-message.ts
