+++ packages/starknet-snap/src/utils/formatter-utils.ts
