+++ packages/starknet-snap/src/utils/index.ts
@@ -4,7 +4,7 @@ export * from './snap';
 export * from './serializer';
 export * from './superstruct';
 export * from './logger';
-export * from './formatterUtils';
+export * from './formatter-utils';
 export * from './snap-state';
 export * from './url';
 export * from './string';
